Basic AI more or less done. Could definitely use more work.
For some reason, units teleport when you tell them to walk back to the
tile they started on. Not sure what's causing this.

Also, units move around on the map editor. To fix this, next time, I'll
have starting a test game make copies of the units onto the new map.